feat: users CRUD RBAC

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\User\UpdateUserRequest;
 use App\Models\User;
 use App\Services\UserService;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 use Inertia\Response;
 
@@ -16,6 +17,8 @@ class UserController extends Controller
 
     public function index(): Response
     {
+        Gate::authorize(ability: 'view-any', arguments: User::class);
+
         return Inertia::render(component: 'users/index', props: [
             'users' => Inertia::scroll(fn () => $this->userService->getAll()),
         ]);
@@ -23,11 +26,15 @@ class UserController extends Controller
 
     public function create(): Response
     {
+        Gate::authorize(ability: 'create', arguments: User::class);
+
         return Inertia::render(component: 'users/create');
     }
 
     public function store(StoreUserRequest $request): RedirectResponse
     {
+        Gate::authorize(ability: 'create', arguments: User::class);
+
         $user = $this->userService->save($request->validated());
         Inertia::flash('toast', ['type' => 'success', 'message' => __(':model :name created successfully', ['model' => __('User'), 'name' => $user->name])]);
 
@@ -36,11 +43,15 @@ class UserController extends Controller
 
     public function edit(User $user): Response
     {
+        Gate::authorize(ability: 'update', arguments: $user);
+
         return Inertia::render(component: 'users/edit', props: ['user' => $user]);
     }
 
     public function update(UpdateUserRequest $request, User $user): RedirectResponse
     {
+        Gate::authorize(ability: 'update', arguments: $user);
+
         $user = $this->userService->update($user, $request->validated());
         Inertia::flash('toast', ['type' => 'success', 'message' => __(':model :name edited successfully', ['model' => __('User'), 'name' => $user->name])]);
 
@@ -49,6 +60,8 @@ class UserController extends Controller
 
     public function destroy(User $user): RedirectResponse
     {
+        Gate::authorize(ability: 'delete', arguments: $user);
+
         $user = $this->userService->delete($user);
         if ($user->trashed()) {
             Inertia::flash('toast', ['type' => 'success', 'message' => __(':model :name deleted successfully', ['model' => __('User'), 'name' => $user->name])]);
@@ -61,6 +74,8 @@ class UserController extends Controller
 
     public function restore(User $user): RedirectResponse
     {
+        Gate::authorize(ability: 'restore', arguments: $user);
+
         $user = $this->userService->restore($user);
         if (! $user->trashed()) {
             Inertia::flash('toast', ['type' => 'success', 'message' => __(':model :name restored successfully', ['model' => __('User'), 'name' => $user->name])]);
